refactor(phase-2i): move execute_large_query_to_json_file to RailsRunnerService
Closes the openproject_client.py god-class decomposition for Rails
execution: ``execute_large_query_to_json_file`` (~241 LOC) is the last
giant Rails-execute method, used by migrations that pull large result
sets. Moving it to ``OpenProjectRailsRunnerService`` collects every
``execute*`` shape on one focused service, alongside
``execute_script_with_data`` (Phase 2h) and the smaller execute family
(Phase 2e–g). ``OpenProjectClient`` keeps a thin delegator so existing
callers work unchanged.

Notes on the move
-----------------
- ``self._check_console_output_for_errors`` calls inside the moved code
  rewire to ``self.check_console_output_for_errors`` (the in-service,
  non-underscored helper from Phase 2e). The client-side underscored
  delegator stays for any external callers but the moved method now
  uses the direct in-service helper.
- ``escape_ruby_single_quoted`` is lazy-imported from
  ``openproject_client`` to avoid the module cycle.
- ``self._client.openproject_rails_runner_service.py`` is added to the
  ``_caller_hint`` skip list so provenance hints continue to point at
  the calling migration, not the runner.
- ``logger`` (module global on the old client) becomes ``self._logger``
  (LoggerAdapter on the service) — same instance via composition.
- Updated the service module docstring header to reflect Phase 2h+2i
  having joined; the heavyweights are no longer "stay on the client".

Numbers
-------
- ``openproject_client.py``: 5,601 → 5,375 LOC (-226)
- ``openproject_rails_runner_service.py``: 877 → 1,128 LOC (+251)
- Cumulative across phases 2a–2i: ``openproject_client.py``
  7,342 → 5,375 LOC (-1,967, -26.8%)

Verification
------------
- pytest tests/unit: 953 passed
- mypy src/: clean (101 files)
- ruff check / ruff format: clean
